requirements.txt
openai>=1.0.0
pyaudio>=0.2.11
speechrecognition>=3.10.0
pyttsx3>=2.90
python-dotenv>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6